requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
fastapi
uvicorn
google-genai
pydantic
psycopg2-binary
passlib[bcrypt]
pydantic[email]
email-validator
